// typescript/models/SearchResponse.ts
/**
 * GameBrain API
 * GameBrain API
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { SearchResponseActiveFilterOptionsInner } from '../models/SearchResponseActiveFilterOptionsInner';
import { SearchResponseFilterOptionsInner } from '../models/SearchResponseFilterOptionsInner';
import { SearchResponseResultsInner } from '../models/SearchResponseResultsInner';
import { SearchResponseSorting } from '../models/SearchResponseSorting';
import { SearchResponseSortingOptionsInner } from '../models/SearchResponseSortingOptionsInner';
import { HttpFile } from '../http/http';

export class SearchResponse {
    'sorting'?: SearchResponseSorting;
    'activeFilterOptions'?: Array<SearchResponseActiveFilterOptionsInner>;
    'query'?: string;
    'totalResults'?: number;
    'limit'?: number;
    'offset'?: number;
    'results'?: Array<SearchResponseResultsInner>;
    'filterOptions'?: Array<SearchResponseFilterOptionsInner>;
    'sortingOptions'?: Array<SearchResponseSortingOptionsInner>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "sorting",
            "baseName": "sorting",
            "type": "SearchResponseSorting",
            "format": ""
        },
        {
            "name": "activeFilterOptions",
            "baseName": "active_filter_options",
            "type": "Array<SearchResponseActiveFilterOptionsInner>",
            "format": ""
        },
        {
            "name": "query",
            "baseName": "query",
            "type": "string",
            "format": ""
        },
        {
            "name": "totalResults",
            "baseName": "total_results",
            "type": "number",
            "format": ""
        },
        {
            "name": "limit",
            "baseName": "limit",
            "type": "number",
            "format": ""
        },
        {
            "name": "offset",
            "baseName": "offset",
            "type": "number",
            "format": ""
        },
        {
            "name": "results",
            "baseName": "results",
            "type": "Array<SearchResponseResultsInner>",
            "format": ""
        },
        {
            "name": "filterOptions",
            "baseName": "filter_options",
            "type": "Array<SearchResponseFilterOptionsInner>",
            "format": ""
        },
        {
            "name": "sortingOptions",
            "baseName": "sorting_options",
            "type": "Array<SearchResponseSortingOptionsInner>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return SearchResponse.attributeTypeMap;
    }

    public constructor() {
    }
}